chore: Remove service-worker.d.ts and update package.json dependencies
The service-worker.d.ts file was deleted because it's not needed anymore.
Also, the versions of several dependencies in package.json were updated to the more recent versions, including typescript, nextjs, and testing libraries.
The upgraded libraries come with bug fixes and performance improvements necessary for the proper functioning and maintenance of the app.